requirements.txt
aiogram==2.22.2
python-dotenv==0.19.2
emoji==2.2.0
